requirements.txt
gTTS
openai
protobuf
pvporcupine
PyAudio
google-cloud-texttospeech
python-dotenv
